+++ lib/src/user/presentation/widgets/reauthenticate_user_dialog.dart
